Reordered some functions and proper resizing
Reordered functions in the file to be easier to read and cluster
together those that deal with similar things. Also changed it so that
resizing the window doesn't redraw everything each event but rather adds
a delay so that all resize events happening within 500ms will delay the
redrawing further. This functionality should be generalized to allow
other keys to be throttled.